Rename container state to instance state
